Spike - move gtm-utils rake task into ga4-faker
Apologies for the giant commit. I will probably re-do this on another branch.

In the gtm-utils repo I refactored the test and fake code and called it from a Rake task.
In this commit I have moved that code into this application. I made a few tweaks as it
moved from a Ruby app to a Rails app.

Final step of the spike will be to see if I can consolidate the Fake code with the faker_job code and call
the class from the job.